TL:DR I cleaned some code and wrote some TODO's
I asked Corey to help center the text for our error messages in our Grid
view because it super bothered me and that help also came along with
some code indentation clean up lol.  While I was there with the project
open I also went through and deleted all the Table stuff and replaced
the simple stuff (Error messages) with Gridview code. I also cleaned up
a bit of the databinding stuff cause we had some repetitive binding. I
wrote a bunch of TODO list items in the code which we apparently can
search in VS by going to TaskList (This was new info to me, hence why I
was so excited to go through and write a bunch of TODO's. Ignore this if
you already knew this ;P)