AddStudent Fixes/Updates
So our problem was once again trying to use multiple entity context
connections, which isn't allowed, so now we are passing the entitiy
context over to the get course offering method likee so:
GetCourseOfferingFromDropdown(entityContext), so that all access stays
in the one context.

GridVIew table now shows, all we needed was a datasource which I've
added in our pre prender but I still couldn't get the GV to function
fully. The error message output was in the code backwards which is why
the wrong error message would show on the GV but I flipped them so it
should be good now.